requirements.txt
aiofile==3.5.0
aiohttp==3.7.4.post0
aioredis==1.3.1
aliyun-log-python-sdk==0.6.53
aliyun-python-sdk-core==2.13.35
aliyun-python-sdk-kms==2.14.0
async-timeout==3.0.1
asyncoss==0.0.4
attrs==21.2.0
caio==0.8.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
crcmod==1.7
cryptography==3.4.7
dateparser==1.0.0
elasticsearch==7.12.1
fastapi==0.63.0
h11==0.12.0
hiredis==2.0.0
httptools==0.1.2
idna==2.10
jmespath==0.10.0
multidict==5.1.0
oss2==2.14.0
paho-mqtt==1.5.1
protobuf==3.17.0
pycparser==2.20
pycryptodome==3.10.1
pydantic==1.8.1
python-dateutil==2.8.1
python-dotenv==0.17.1
python-multipart==0.0.5
pytz==2021.1
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
six==1.16.0
starlette==0.13.6
typing-extensions==3.10.0.0
tzlocal==2.1
urllib3==1.26.4
uvicorn==0.13.4
uvloop==0.15.2
watchgod==0.7
websockets==8.1
yarl==1.6.3
tortoise-orm==0.17.7
aiofiles==0.8.0
xlrd==2.0.1
openpyxl==3.0.7
APScheduler==3.8.1
